Rework into src folder to dogfood builds
