+++ prettier.config.mjs
@@ -1,14 +1,1 @@
-/** @type {import('prettier').Config} */
-const config = {
-  tabWidth: 2,
-  useTabs: false,
-  semi: true,
-  singleQuote: true,
-  trailingComma: 'es5',
-  bracketSpacing: true,
-  arrowParens: 'always',
-  proseWrap: 'never',
-  printWidth: 80,
-};
-
-export default config;
+export { default } from './dist/prettier.config.mjs';
