+++ src/rules/compat.mjs
