+++ src/rules/eslint-typescript.mjs
