+++ src/rules/eslint.mjs
