+++ src/rules/n.mjs
