+++ src/rules/promise.mjs
