+++ src/rules/typescript.mjs
