+++ tsconfig.json
@@ -15,6 +15,6 @@
     "jsx": "react-jsx",
     "noEmit": true
   },
-  "include": ["src", "test", "tsup.config.ts", "*.config.mjs"],
+  "include": ["src", "test", "tsup.config.ts"],
   "exclude": ["node_modules", "dist"]
 }
